requirements.txt
lxml>=4.9.3
requests>=2.31.0
Twisted[tls]>=22.10.0
zope.interface>=6.0
